+++ portal/tests/test_api.py
@@ -8,7 +8,6 @@ from common.models import Student, Teacher
 from common.tests.utils.user import create_user_directly, get_superuser
 from django.contrib.auth.models import User
 from django.urls import reverse
-from django.utils import timezone
 from hamcrest import *
 from hamcrest.core.base_matcher import BaseMatcher
 from rest_framework import status
@@ -70,7 +69,7 @@ class APITests(APITestCase):
         client.force_authenticate(user=superuser)
         url = reverse("inactive_users")
         response = client.get(url)
-        self.assertEqual(len(response.data), 1)
+        assert len(response.data) == 1
 
     @patch("portal.views.api.IS_CLOUD_SCHEDULER_FUNCTION", return_value=True)
     def test_get_inactive_users_if_appengine(self, mock_is_cloud_scheduler_function):
@@ -89,7 +88,7 @@ class APITests(APITestCase):
         create_user_directly(active=True)
         url = reverse("inactive_users")
         response = client.get(url)
-        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
+        assert response.status_code == status.HTTP_403_FORBIDDEN
 
     @patch("portal.views.api.IS_CLOUD_SCHEDULER_FUNCTION", return_value=True)
     def test_delete_inactive_users_if_appengine(self, mock_is_cloud_scheduler_function):