+++ portal/views/home.py
@@ -1,13 +1,17 @@
+import logging
+
 from common import email_messages
 from common.helpers.emails import (
     send_verification_email,
     send_email,
     NOTIFICATION_EMAIL,
     add_to_dotmailer,
+    DotmailerUserType,
 )
 from common.models import Teacher, Student
 from common.permissions import logged_in_as_student, logged_in_as_teacher
 from common.utils import _using_two_factor
+from deploy import captcha
 from django.contrib import messages as messages
 from django.contrib.auth import logout
 from django.contrib.auth.models import User
@@ -15,12 +19,17 @@ from django.http import HttpResponseRedirect
 from django.shortcuts import render
 from django.urls import reverse, reverse_lazy
 from django.views.decorators.cache import cache_control
-
-from deploy import captcha
 from portal.forms.play import IndependentStudentSignupForm
 from portal.forms.teach import TeacherSignupForm
 from portal.helpers.captcha import remove_captcha_from_forms
+from portal.helpers.ratelimit import (
+    RATELIMIT_USER_ALREADY_REGISTERED_EMAIL_GROUP,
+    RATELIMIT_USER_ALREADY_REGISTERED_EMAIL_RATE,
+    is_ratelimited,
+)
 from portal.strings.home_learning import HOME_LEARNING_BANNER
+
+LOGGER = logging.getLogger(__name__)
 
 
 def teach_email_labeller(request):
@@ -114,12 +123,25 @@ def process_signup_form(request, data):
 
     if email and User.objects.filter(email=email).exists():
         email_message = email_messages.userAlreadyRegisteredEmail(request, email)
-        send_email(
-            NOTIFICATION_EMAIL,
-            [email],
-            email_message["subject"],
-            email_message["message"],
+        is_email_ratelimited = is_ratelimited(
+            request=request,
+            group=RATELIMIT_USER_ALREADY_REGISTERED_EMAIL_GROUP,
+            key=lambda *_: email,
+            rate=RATELIMIT_USER_ALREADY_REGISTERED_EMAIL_RATE,
+            increment=True,
         )
+
+        if not is_email_ratelimited:
+            send_email(
+                NOTIFICATION_EMAIL,
+                [email],
+                email_message["subject"],
+                email_message["message"],
+            )
+        else:
+            LOGGER.warn(
+                f"Ratelimit teacher {RATELIMIT_USER_ALREADY_REGISTERED_EMAIL_GROUP}: {email}"
+            )
     else:
         teacher = Teacher.objects.factory(
             first_name=data["teacher_first_name"],
@@ -130,7 +152,9 @@ def process_signup_form(request, data):
 
         if _newsletter_ticked(data):
             user = teacher.user.user
-            add_to_dotmailer(user.first_name, user.last_name, user.email)
+            add_to_dotmailer(
+                user.first_name, user.last_name, user.email, DotmailerUserType.TEACHER
+            )
 
         send_verification_email(request, teacher.user.user)
 
@@ -149,12 +173,25 @@ def process_independent_student_signup_form(request, data):
         email_message = email_messages.userAlreadyRegisteredEmail(
             request, email, is_independent_student=True
         )
-        send_email(
-            NOTIFICATION_EMAIL,
-            [email],
-            email_message["subject"],
-            email_message["message"],
+        is_email_ratelimited = is_ratelimited(
+            request=request,
+            group=RATELIMIT_USER_ALREADY_REGISTERED_EMAIL_GROUP,
+            key=lambda *_: email,
+            rate=RATELIMIT_USER_ALREADY_REGISTERED_EMAIL_RATE,
+            increment=True,
         )
+
+        if not is_email_ratelimited:
+            send_email(
+                NOTIFICATION_EMAIL,
+                [email],
+                email_message["subject"],
+                email_message["message"],
+            )
+        else:
+            LOGGER.warn(
+                f"Ratelimit independent {RATELIMIT_USER_ALREADY_REGISTERED_EMAIL_GROUP}: {email}"
+            )
         return render(
             request,
             "portal/email_verification_needed.html",
@@ -170,7 +207,9 @@ def process_independent_student_signup_form(request, data):
 
     if _newsletter_ticked(data):
         user = student.new_user
-        add_to_dotmailer(user.first_name, user.last_name, user.email)
+        add_to_dotmailer(
+            user.first_name, user.last_name, user.email, DotmailerUserType.STUDENT
+        )
 
     send_verification_email(request, student.new_user)