+++ portal_frontend/src/ThemeTypes.d.ts
@@ -1,0 +1,11 @@
+import React from "react";
+
+declare module "@mui/material/styles" {
+  // allow configuration using `createTheme`
+  interface PaletteOptions {
+    tertiary?: PaletteColorOptions;
+  }
+  interface TypographyVariantsOptions {
+    quote?: React.CSSProperties;
+  }
+}